requirements.txt
# python: 3.10.14
pyspark==3.5.1
pyyaml==6.0.2
pytest==8.4.1
pytest-cov==5.0.0
pydantic==2.5.0
pandas==2.2.3
pyarrow==11.0.0
numpy==1.26.4